added comment rest api #343

+++ app/models/Comment.java
@@ -3,12 +3,17 @@ package models;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.typesafe.plugin.RedisPlugin;
 import org.joda.time.DateTime;
 import play.Logger;
 import play.libs.Json;
+import redis.clients.jedis.Jedis;
 
 import javax.persistence.*;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 @Entity
 public class Comment extends SuperModel {
@@ -66,21 +71,31 @@ public class Comment extends SuperModel {
 		rs.refType = refType;
 		rs.save();
 
+		// add to cache
+		rs.hmset();
+		rs.zadd(refId.toString(), refType);
+
 		return rs;
 	}
 
 	/*
-	Remove comment from an object identified by refId and refType
+	Remove comment from an object identified by refId and refType and check if user owns it
 	 */
-	public static void removeComment(Long commentId, Long refId, String refType) {
+	public static boolean deleteComment(Long commentId, Long refId, String refType, Long userId) {
 		Comment rs = null;
 
 		rs = Comment.find.byId(commentId);
 
-		// if found then delete
-		if(rs != null){
+		// if found then delete if user owns it
+		if(rs != null && rs.user.id == userId){
+			// delete from cache
+			rs.zrem(refId.toString(), refType);
+			rs.del();
+			// delete from db
 			rs.delete();
+			return true;
 		}
+		return false;
 	}
 
 	/*
@@ -133,5 +148,117 @@ public class Comment extends SuperModel {
 			arrayNode.add(item.toJson());
 		}
 		return arrayNode;
+	}
+
+	public void hmset(){
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		try {
+			Map hashmap = new HashMap();
+
+			// Create the hashmap values
+			hashmap.put("id", this.id.toString());
+			hashmap.put("createdAt", new DateTime(this.createdAt).toString());
+			hashmap.put("body", this.body);
+
+			if(this.user != null) {
+				hashmap.put("user.id", this.user.id.toString());
+			}
+
+			// add the values
+			j.hmset("comment:" + this.id.toString(), hashmap);
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+	}
+
+	public static ObjectNode hmget(String id){
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		ObjectNode node = Json.newObject();
+		try {
+			String key = "comment:" + id.toString();
+
+			if(!j.exists(key)) {
+				Logger.debug("comment added to cache " + key);
+				Comment.find.byId(Long.parseLong(id, 10)).hmset();
+			}
+
+			List<String> values = j.hmget(key, "id", "createdAt", "body", "user.id");
+
+			if (values.get(0) != null) node.put("id", values.get(0));
+			if (values.get(1) != null) node.put("createdAt", values.get(1));
+			if (values.get(2) != null) node.put("body", values.get(2));
+			if (values.get(3) != null) {
+				node.put("user", User.hmget(values.get(3)));
+			}
+			node.put("likes", Likes.range(id, "models.comment"));
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+		return node;
+	}
+
+	public void del(){
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		try {
+			// delete
+			j.del("comment:" + this.id.toString());
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+	}
+
+	public void zadd(String refId, String refType){
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		try {
+			// compose key
+			String key = "comments:"+ refType + ":" + refId;
+			// left push to list
+			Logger.debug("comment added to list " + key);
+			j.zadd(key, createdAt.getTime(), this.id.toString());
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+	}
+
+	public void zrem(String refId, String refType){
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		try {
+			// compose key
+			String key = "comments:"+ refType + ":" + refId;
+			// delete
+			j.zrem(key, this.id.toString());
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+	}
+
+	public static ObjectNode range(String refId, String refType){
+		ObjectNode items = Json.newObject();
+		ArrayNode data = items.putArray("data");
+
+		//Go to Redis to read the full roster of content.
+		Jedis j = play.Play.application().plugin(RedisPlugin.class).jedisPool().getResource();
+		try {
+			String key = "comments:" + refType + ":" + refId;
+
+			if(!j.exists(key)) {
+				Logger.debug("adding comments to cache: " + key);
+				for(Comment item: getAllComments(Long.parseLong(refId, 10), refType)){
+					item.zadd(refId, refType);
+				}
+			}
+
+			Set<String> set = j.zrange(key, 0, -1);
+			items.put("count", j.zcard(key).toString());
+
+			for(String id: set) {
+				// get the data for each like
+				data.add(Comment.hmget(id));
+			}
+		} finally {
+			play.Play.application().plugin(RedisPlugin.class).jedisPool().returnResource(j);
+		}
+
+		return items;
 	}
 }
